third_party/blink/web_tests/flag-specific/highdpi/printing/css2.1/page-break-after-003-expected.txt

There must not be a page break after this line of text.

This test should produce two pages of output.
This text should be at the top of the second and final page.


PASS: page number of "test1" is 0
FAIL: expected page number of "test2" is 1. Was 0
FAIL: expected number of pages is 2. Was 1
